+++ Packages/com.unity.inputsystem/InputSystem/Devices/Joystick.cs
@@ -106,7 +106,22 @@ namespace UnityEngine.InputSystem
         /// See <see cref="InputDevice.MakeCurrent"/> for details about when a device
         /// is made current.
         /// </remarks>
+        /// <seealso cref="all"/>
         public static Joystick current { get; private set; }
+
+        /// <summary>
+        /// A list of joysticks currently connected to the system.
+        /// </summary>
+        /// <value>All currently connected joystick.</value>
+        /// <remarks>
+        /// Does not cause GC allocation.
+        ///
+        /// Do <em>not</em> hold on to the value returned by this getter but rather query it whenever
+        /// you need it. Whenever the joystick setup changes, the value returned by this getter
+        /// is invalidated.
+        /// </remarks>
+        /// <seealso cref="current"/>
+        public new static ReadOnlyArray<Joystick> all => new ReadOnlyArray<Joystick>(s_Joysticks, 0, s_JoystickCount);
 
         /// <summary>
         /// Called when the joystick has been created but before it is added
@@ -140,13 +155,35 @@ namespace UnityEngine.InputSystem
         }
 
         /// <summary>
+        /// Called when the joystick is added to the system.
+        /// </summary>
+        protected override void OnAdded()
+        {
+            ArrayHelpers.AppendWithCapacity(ref s_Joysticks, ref s_JoystickCount, this);
+        }
+
+        /// <summary>
         /// Called when the joystick is removed from the system.
         /// </summary>
         protected override void OnRemoved()
         {
             base.OnRemoved();
+
             if (current == this)
                 current = null;
+
+            // Remove from `all`.
+            var index = ArrayHelpers.IndexOfReference(s_Joysticks, this, s_JoystickCount);
+            if (index != -1)
+                ArrayHelpers.EraseAtWithCapacity(s_Joysticks, ref s_JoystickCount, index);
+            else
+            {
+                Debug.Assert(false,
+                    $"Joystick {this} seems to not have been added but is being removed (joystick list: {string.Join(", ", all)})"); // Put in else to not allocate on normal path.
+            }
         }
+
+        private static int s_JoystickCount;
+        private static Joystick[] s_Joysticks;
     }
 }